motor_controller/requirements.txt
RPi.GPIO==0.7.0
python-dotenv==0.19.0
json-rpc==1.13.0